controller angepasst

+++ src/main/java/de/SetMyRoute/controller/DifficultyController.java
@@ -1,5 +1,6 @@
 package de.SetMyRoute.controller;
 
+import de.SetMyRoute.model.User;
 import de.SetMyRoute.model.difficulty.Difficulty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -13,18 +14,33 @@ public class DifficultyController
     @Autowired
     public DifficultyRepository difficultyRepository;
 
-    @DeleteMapping("/difficulty/delete/{id}")
+    @Autowired
+    public UserRepository userRepository;
+
+    @DeleteMapping("/difficulty/delete/{id}/{authToken}")
     public @ResponseBody
-    String deleteDifficulty(@PathVariable int id) {
-        difficultyRepository.deleteById(id);
-        return "Deleted";
+    String deleteDifficulty(@PathVariable int id, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                difficultyRepository.deleteById(id);
+                return "Deleted";
+            }
+        }
+        return "Not allowed!";
     }
 
-    @PostMapping(path = "/difficulty")
+    @PostMapping(path = "/difficulty/{authToken}")
     public @ResponseBody
-    String addDifficulty(Difficulty difficulty) {
-        difficultyRepository.save(difficulty);
-        return "added";
+    String addDifficulty(Difficulty difficulty, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                difficultyRepository.save(difficulty);
+                return "added";
+            }
+        }
+        return "Not allowed!";
     }
 
     @GetMapping(path = "/difficulty/{difficultyId}")
@@ -33,7 +49,7 @@ public class DifficultyController
         return difficultyRepository.findById(difficultyId);
     }
 
-    @GetMapping(path = "/difficulty")
+    @GetMapping(path = "/difficultys")
     public @ResponseBody
     Iterable<Difficulty> findAllDifficultys() {
         return difficultyRepository.findAll();
