+++ src/main/java/de/SetMyRoute/controller/HoldController.java
@@ -1,6 +1,7 @@
 package de.SetMyRoute.controller;
 
 import de.SetMyRoute.model.Hold;
+import de.SetMyRoute.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,19 +14,34 @@ public class HoldController {
     @Autowired
     public HoldRepository holdRepository;
 
+    @Autowired
+    public UserRepository userRepository;
 
-    @DeleteMapping("/hold/delete/{holdId}")
+
+    @DeleteMapping("/hold/delete/{holdId}/{authToken}")
     public @ResponseBody
-    String deleteHold(@PathVariable int holdId) {
-        holdRepository.deleteById(holdId);
-        return "Deleted";
+    String deleteHold(@PathVariable int holdId, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                holdRepository.deleteById(holdId);
+                return "Deleted";
+            }
+        }
+        return "Not allowed!";
     }
 
-    @PostMapping(path = "/hold")
+    @PostMapping(path = "/hold/{authToken}")
     public @ResponseBody
-    String addHold(Hold hold) {
-        holdRepository.save(hold);
-        return "added";
+    String addHold(Hold hold, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                holdRepository.save(hold);
+                return "added";
+            }
+        }
+        return "Not allowed!";
     }
 
     @GetMapping(path = "/hold/{holdId}")
