+++ src/main/java/de/SetMyRoute/controller/RouteController.java
@@ -15,6 +15,9 @@ public class RouteController {
     @Autowired
     public RouteRepository routeRepository;
 
+    @Autowired
+    public UserRepository userRepository;
+
     @GetMapping(path = "/route/{routeId}")
     public @ResponseBody
     Optional<Route> getRouteById(@PathVariable Integer routeId) {
@@ -27,17 +30,29 @@ public class RouteController {
         return routeRepository.findAll();
     }
 
-    @DeleteMapping("/route/{routeId}")
+    @DeleteMapping("/route/{routeId}/{authToken}")
     public @ResponseBody
-    String deleteRoute(@PathVariable Integer routeId) {
-        routeRepository.deleteById(routeId);
-        return "Deleted";
+    String deleteRoute(@PathVariable Integer routeId, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                routeRepository.deleteById(routeId);
+                return "Deleted";
+            }
+        }
+        return "Not allowed!";
     }
 
-    @PostMapping(path = "/route")
+    @PostMapping(path = "/route/{authToken}")
     public @ResponseBody
-    String addRoute(Route route) {
-        routeRepository.save(route);
-        return "added";
+    String addRoute(Route route, @PathVariable String authToken) {
+        Optional<User> user = userRepository.findByAuthToken(authToken);
+        if(user.isPresent()) {
+            if (user.get().isAdmin()) {
+                routeRepository.save(route);
+                return "added";
+            }
+        }
+        return "Error 400";
     }
 }
